+++ ouroboros-network/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking changes
 
 ### Non-Breaking changes
+* Honour policyPeerShareActivationDelay timeout when peersharing
 
 ## 0.12.0.0 -- 2023-02-21